Move everything into similar-proteins directory.
