+++ src/uniprotkb/components/entry/similar-proteins/AgrOrthology.tsx
@@ -1,17 +1,17 @@
 import { ExternalLink, Loader } from 'franklin-sites';
 import { Link } from 'react-router-dom';
 
-import { Location, LocationToPath } from '../../../app/config/urls';
-import ErrorHandler from '../../../shared/components/error-pages/ErrorHandler';
+import { Location, LocationToPath } from '../../../../app/config/urls';
+import ErrorHandler from '../../../../shared/components/error-pages/ErrorHandler';
 import TableFromData, {
   TableFromDataColumn,
-} from '../../../shared/components/table/TableFromData';
-import WithTooltip from '../../../shared/components/WithTooltip';
-import externalUrls from '../../../shared/config/externalUrls';
-import useDataApi from '../../../shared/hooks/useDataApi';
-import * as logging from '../../../shared/utils/logging';
-import { stringifyQuery } from '../../../shared/utils/url';
-import { AgrOrthologs, AgrOrthologsResult } from '../../types/agrOrthologs';
+} from '../../../../shared/components/table/TableFromData';
+import WithTooltip from '../../../../shared/components/WithTooltip';
+import externalUrls from '../../../../shared/config/externalUrls';
+import useDataApi from '../../../../shared/hooks/useDataApi';
+import * as logging from '../../../../shared/utils/logging';
+import { stringifyQuery } from '../../../../shared/utils/url';
+import { AgrOrthologs, AgrOrthologsResult } from '../../../types/agrOrthologs';
 import styles from './styles/agr-homology.module.scss';
 
 // Lifted from https://github.com/alliance-genome/agr_ui/blob/6f5acc104df6274bb0642a2317a5b6b102a91b32/src/components/orthology/orthologyTable.js#L29
