+++ src/uniprotkb/components/entry/similar-proteins/styles/agr-homology.module.scss
